Refactor lowerer_impl to be abstract with a lowerer_impl_C version
